refactor(domain): split timetable-utils by concern
Replace the grab-bag timetable-utils.ts with three modules, each
named after the responsibility it owns:

- timetable-entry-boarding.ts: boarding/alighting predicates over a
  single entry (isDropOffOnly, isBoardingOnly, requiresArrangement,
  isPassThrough). This is where the Issue #162 state-model work will
  land; the file is anchored to the stable entry.boarding field name
  rather than the judgment vocabulary that #162 may rename.
- timetable-entry-schedule.ts: schedule reading helpers
  (getDisplayMinutes, hasDwellTime, getDwellMinutes,
  getRemainingMinutes).
- timetable-service-state.ts: service-state derivation
  (getStopServiceState, getTimetableEntriesState,
  getFilteredTimetableEntriesState) plus hasBoardable, the
  exists-lift bridging entry predicates to collection state.

Function bodies and TSDoc move verbatim; only module docs are new.
Update the 18 import sites, split the test file accordingly, and
extract the shared makeEntry/testRoute fixture into
__tests__/make-timetable-entry.ts instead of duplicating it. The
TimetableEntry TSDoc path references now point to
timetable-entry-boarding.ts. No behavior change.

Refs: #162
