+++ src/repositories/athenai-repository/athenai-repository-v2.ts
@@ -35,7 +35,7 @@ import {
   sortTimetableEntriesByDepartureTime,
   sortTimetableEntriesChronologically,
 } from '../../domain/transit/sort-timetable-entries';
-import { getTimetableEntriesState } from '../../domain/transit/timetable-utils';
+import { getTimetableEntriesState } from '../../domain/transit/timetable-service-state';
 import { createLogger } from '../../lib/logger';
 import type { Bounds, LatLng, RouteShape } from '../../types/app/map';
 import type {
